+++ client/src/components/Home/Rich/index.js
