+++ client/src/components/Home/index.js
@@ -1,0 +1,13 @@
+// Import
+
+// From Rich Folder 
+import Pie from "./Rich/Pie";
+import Chart from "./Rich/Chart";
+import RichList from "./Rich/RichList";
+
+// Export 
+export {
+  Pie,
+  Chart,
+  RichList,
+};
